Moved the code to support test threaded socket servers into its own module.
